+++ src/mem/vmem/vmem.c
@@ -31,9 +31,9 @@ extern char _text_vma, _rodata_vma, _data_vma, _bss_vma;
 extern char _text_len, _rodata_len, _data_len, _bss_len_with_reserve;
 
 void vmem_get_idx_from_vaddr(mmu_vaddr_t virt_addr, size_t *pgd_idx, size_t *pmd_idx, size_t *pte_idx) {
-	*pgd_idx = ((uint32_t) virt_addr & MMU_PGD_MASK) >> MMU_PGD_SHIFT;
-	*pmd_idx = ((uint32_t) virt_addr & MMU_PMD_MASK) >> MMU_PMD_SHIFT;
-	*pte_idx = ((uint32_t) virt_addr & MMU_PTE_MASK) >> MMU_PTE_SHIFT;
+	*pgd_idx = (size_t) (virt_addr & MMU_PGD_MASK) >> MMU_PGD_SHIFT;
+	*pmd_idx = (size_t) (virt_addr & MMU_PMD_MASK) >> MMU_PMD_SHIFT;
+	*pte_idx = (size_t) (virt_addr & MMU_PTE_MASK) >> MMU_PTE_SHIFT;
 }
 
 static int vmem_kernel_map(void *start, uint32_t len, uint32_t flags) {
